feat(alerts): update alert creation and validation to include default values and camelCase fields feat(stocks): refine stock validation to ensure positive values for volume and market cap feat(stockPrice): change volume field type to DECIMAL for better precision

+++ src/database/models/stockPrice.model.js
@@ -57,7 +57,7 @@ module.exports = (sequelize) => {
       field: 'adjusted_close',
     },
     volume: {
-      type: DataTypes.BIGINT,
+      type: DataTypes.DECIMAL(15, 4),
       allowNull: true,
     },
     dataSource: {
